change strong item check timing(6569)
