+++ nekoyume/Assets/_Scripts/UI/Widget/Popup/SynthesisRegistrationPopup.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using Cysharp.Threading.Tasks;
 using Nekoyume.Game.Controller;
 using Nekoyume.L10n;
+using Nekoyume.Model.Item;
 using Nekoyume.Model.Mail;
 using Nekoyume.UI.Model;
 using Nekoyume.UI.Module;
@@ -136,6 +137,7 @@ namespace Nekoyume.UI
                 return;
             }
 
+            // 선택된 아이템이 필요 수량으로 나누어 떨어지는지 확인
             var selectedCount = synthesisInventory.SelectedItems.Count;
             if (selectedCount % _synthesizeModel.RequiredItemCount != 0)
             {
@@ -143,6 +145,7 @@ namespace Nekoyume.UI
                 return;
             }
 
+            // 최대 합성 가능 수량 확인
             var synthesisCount = selectedCount / _synthesizeModel.RequiredItemCount;
             if (synthesisCount > Synthesis.MaxSynthesisCount)
             {
@@ -150,6 +153,7 @@ namespace Nekoyume.UI
                 return;
             }
 
+            // 장착된 장비가 포함되어 있는지 확인
             foreach (var selectedItem in synthesisInventory.SelectedItems)
             {
                 if (!selectedItem.Equipped.Value)
@@ -165,7 +169,35 @@ namespace Nekoyume.UI
                 return;
             }
 
+            // 강화된 장비가 포함되어 있는지 확인
+            foreach (var selectedItem in synthesisInventory.SelectedItems)
+            {
+                if (!Synthesis.IsStrong(selectedItem.ItemBase))
+                {
+                    continue;
+                }
+
+                var system = Widget.Find<IconAndButtonSystem>();
+                system.ShowWithTwoButton("UI_WARNING", "UI_SYNTHESIZE_STRONG_CONFIRM");
+                system.ConfirmCallback = () => RegisterItem(synthesisInventory.SelectedItems);
+                return;
+            }
+
             RegisterItem(synthesisInventory.SelectedItems);
+        }
+
+        private void CheckSynthesizeStringEquipment(List<ItemBase> itemBaseList, System.Action callback)
+        {
+            if (itemBaseList.Exists(Synthesis.IsStrong))
+            {
+                var system = Widget.Find<IconAndButtonSystem>();
+                system.ShowWithTwoButton("UI_WARNING", "UI_SYNTHESIZE_STRONG_CONFIRM");
+                system.ConfirmCallback = callback;
+            }
+            else
+            {
+                callback();
+            }
         }
 
         private void OnClickInventoryItem(InventoryItem item)
@@ -235,8 +267,6 @@ namespace Nekoyume.UI
             }
 
             _registerMaterials.Invoke(items, _synthesizeModel);
-
-            // TODO: 강화된 아이템 체크
             CloseWidget.Invoke();
         }
